api/requirements.txt
fastapi
uvicorn
pymongo[srv]
python-dotenv
pydantic
